+++ tests/testthat/test-detect_same_place.R
@@ -17,7 +17,7 @@
 #  useful, but it comes WITHOUT ANY WARRANTY OR LIABILITY.              #
 # ===================================================================== #
 
-same_place_case <- function(source_key, sample_date, institution_id, pathogen = "Test organism",
+same_place_case <- function(source_key, sample_date, institution_id, pathogen = "Test pathogen",
                              ward = "ICU") {
   data.frame(
     source_key = source_key, pathogen = pathogen, institution_id = institution_id, ward = ward,